Revert "Update to latest Datastream GA library (#65)"
This reverts commit 12ed79f3678ec7f0d353c8cf72d2ff51640158ad.
