+++ src/main/java/io/cdap/delta/datastream/DatastreamTableAssessor.java
@@ -18,14 +18,14 @@ package io.cdap.delta.datastream;
 
 
 import com.google.api.core.ApiFuture;
-import com.google.cloud.datastream.v1.CreateConnectionProfileRequest;
-import com.google.cloud.datastream.v1.CreateStreamRequest;
-import com.google.cloud.datastream.v1.DatastreamClient;
-import com.google.cloud.datastream.v1.OperationMetadata;
-import com.google.cloud.datastream.v1.OracleRdbms;
-import com.google.cloud.datastream.v1.Stream;
-import com.google.cloud.datastream.v1.Validation;
-import com.google.cloud.datastream.v1.ValidationMessage;
+import com.google.cloud.datastream.v1alpha1.CreateConnectionProfileRequest;
+import com.google.cloud.datastream.v1alpha1.CreateStreamRequest;
+import com.google.cloud.datastream.v1alpha1.DatastreamClient;
+import com.google.cloud.datastream.v1alpha1.OperationMetadata;
+import com.google.cloud.datastream.v1alpha1.OracleRdbms;
+import com.google.cloud.datastream.v1alpha1.Stream;
+import com.google.cloud.datastream.v1alpha1.Validation;
+import com.google.cloud.datastream.v1alpha1.ValidationMessage;
 import com.google.cloud.storage.Storage;
 import com.google.cloud.storage.StorageException;
 import com.google.common.annotations.VisibleForTesting;
@@ -191,7 +191,7 @@ public class DatastreamTableAssessor implements TableAssessor<TableDetail> {
                         streamPath, e.getLocalizedMessage()), e);
       }
 
-      OracleRdbms originalAllowList = stream.getSourceConfig().getOracleSourceConfig().getIncludeObjects();
+      OracleRdbms originalAllowList = stream.getSourceConfig().getOracleSourceConfig().getAllowlist();
       Utils.addToAllowList(stream, new HashSet<>(tables));
       try {
         Utils.updateAllowlist(datastream, stream.build(), true, LOGGER);
@@ -318,8 +318,8 @@ public class DatastreamTableAssessor implements TableAssessor<TableDetail> {
     List<Problem> connectivityIssues = new ArrayList<>();
     List<Problem> missingFeatures = new ArrayList<>();
     for (Validation validation : metadata.getValidationResult().getValidationsList()) {
-      Validation.State state = validation.getState();
-      if (Validation.State.FAILED.equals(state)) {
+      Validation.Status status = validation.getStatus();
+      if (Validation.Status.FAILED.equals(status)) {
         String code = validation.getCode();
         String description = validation.getDescription();
         String message = String.join("\n",
